# GenericDataChunkIterator improvements