--- a/tests/test_gin.py
+++ b/tests/test_gin.py
@@ -29,7 +29,7 @@ except ImportError:
 #   ecephys: https://gin.g-node.org/NeuralEnsemble/ephy_testing_data
 #   ophys: TODO
 #   icephys: TODO
-LOCAL_PATH = Path(".")  # Must be set to "." for CI - temporarily override for local testing
+LOCAL_PATH = Path("/home/jovyan")  # Must be set to "." for CI - temporarily override for local testing
 DATA_PATH = LOCAL_PATH / "ephy_testing_data"
 HAVE_DATA = DATA_PATH.exists()
 
@@ -92,7 +92,7 @@ if HAVE_PARAMETERIZED and HAVE_DATA:
 
         @parameterized.expand(parameterized_recording_list)
         def test_convert_recording_extractor_to_nwb(self, recording_interface, interface_kwargs):
-            nwbfile_path = self.savedir / f"{recording_interface.__name__}.nwb"
+            nwbfile_path = str(self.savedir / f"{recording_interface.__name__}.nwb")
 
             class TestConverter(NWBConverter):
                 data_interface_classes = dict(TestRecording=recording_interface)
@@ -105,7 +105,9 @@ if HAVE_PARAMETERIZED and HAVE_DATA:
             check_recordings_equal(RX1=recording, RX2=nwb_recording, check_times=False, return_scaled=True)
             # Technically, check_recordings_equal only tests a snippet of data. Above tests are for metadata mostly.
             # For GIN test data, sizes should be OK to load all into RAM even on CI
-            npt.assert_array_equal(x=recording.get_traces(), y=nwb_recording.get_traces())
+            npt.assert_array_equal(
+                x=recording.get_traces(return_scaled=False), y=nwb_recording.get_traces(return_scaled=False)
+            )
 
         @parameterized.expand(
             [
@@ -120,7 +122,7 @@ if HAVE_PARAMETERIZED and HAVE_DATA:
             ]
         )
         def test_convert_sorting_extractor_to_nwb(self, sorting_interface, interface_kwargs):
-            nwbfile_path = self.savedir / f"{sorting_interface.__name__}.nwb"
+            nwbfile_path = str(self.savedir / f"{sorting_interface.__name__}.nwb")
 
             class TestConverter(NWBConverter):
                 data_interface_classes = dict(TestSorting=sorting_interface)